fixed redirect for display session. Pending fix for start session so that devices publish. Will include start session on display instead
