+++ simuliot/rootfs/simuliot_backend/simuliot.py
@@ -13,6 +13,7 @@ import mqtt_config as mqtt_cfg
 import sqlite3
 import logging
 import os, sys
+import mqtt_client as mqtt_client
 import json
 
 logger = logging.getLogger(__name__)
@@ -38,16 +39,6 @@ def signal_handler(sig, frame):
     logger.info('Shutting down Simulated IoT Device')
     sys.exit(0)
 
-def on_connect(client, userdata, flags, rc):
-    logger.info('connected ({})'.format(client._client_id))
-    client.subscribe(topic=mqtt_cfg.TOPIC, qos=2)
-
-def on_message(client, userdata, message):
-    logger.info('------------------------------')
-    decoded = json.loads(message.payload.decode())
-    logger.info("topic: {}, msg: {}".format(
-        message.topic, decoded))
-
 def connect_db():
     conn = None
     if (os.path.isfile('../tfg-test.db')):
@@ -60,50 +51,37 @@ def connect_db():
         logger.critical('Database file not found. Please reinstall addon.')
     return conn
 
-def connect_mqtt():
-    try:
-        Client = mqtt.Client(callback_api_version=mqtt.CallbackAPIVersion.VERSION2, client_id="SimulIoT Session Client", clean_session=False)
-        Client.on_connect = on_connect
-        Client.on_message = on_message
-        Client.username_pw_set(mqtt_cfg.mqtt_credentials["user"], mqtt_cfg.mqtt_credentials["pwd"])
-        Client.connect(host=mqtt_cfg.mqtt_broker["host"], port=mqtt_cfg.mqtt_broker["port"])
-        return Client
-    except Exception as e:
-        logger.critical('Error: ' + str(e))
-
-client = connect_mqtt()
-
 def getTempDevice(deviceID, deviceName, location, type, isContactProbe):
-    tempDevice = temp.tempDevice(deviceID, deviceName, location, type, client, isContactProbe)
+    tempDevice = temp.tempDevice(deviceID, deviceName, location, type, mqtt_client.client(), isContactProbe)
     return tempDevice
 
 def getSwitchConfigurableDevice(deviceID, deviceName, location, type, config):
     ## All configs will be empty for now
-    switchConfigurableDevice = switch_config.configurableSwitchDevice(deviceID, deviceName, location, type, client, config)
+    switchConfigurableDevice = switch_config.configurableSwitchDevice(deviceID, deviceName, location, type, mqtt_client.client(), config)
     return switchConfigurableDevice
 
 def getSwitchDevice(deviceID, deviceName, location, type):
-    switchDevice = switch.switchDevice(deviceID, deviceName, location, type, client)
+    switchDevice = switch.switchDevice(deviceID, deviceName, location, type, mqtt_client.client())
     return switchDevice
 
 def getFlowDevice(deviceID, deviceName, location, type, kind):
-    flowDevice = flow.flowDevice(deviceID, deviceName, location, type, client, kind)
+    flowDevice = flow.flowDevice(deviceID, deviceName, location, type, mqtt_client.client(), kind)
     return flowDevice
 
 def getTempSwitchDevice(deviceID, deviceName, location, type):
-    tempSwitchDevice = temp_switch.tempSwitchDevice(deviceID, deviceName, location, type, client)
+    tempSwitchDevice = temp_switch.tempSwitchDevice(deviceID, deviceName, location, type, mqtt_client.client())
     return tempSwitchDevice
 
 def getPresenceDevice(deviceID, deviceName, location, type):
-    presenceDevice = presence.presenceDevice(deviceID, deviceName, location, type, client)
+    presenceDevice = presence.presenceDevice(deviceID, deviceName, location, type, mqtt_client.client())
     return presenceDevice
 
 def getSoundSensorDevice(deviceID, deviceName, location, type):
-    soundSensorDevice = sound.soundSensorDevice(deviceID, deviceName, location, type, client)
+    soundSensorDevice = sound.soundSensorDevice(deviceID, deviceName, location, type, mqtt_client.client())
     return soundSensorDevice
 
 def getHubDevice(deviceID, deviceName, location, type):
-    hubDevice = hub.hubDevice(deviceID, deviceName, location, type, client)
+    hubDevice = hub.hubDevice(deviceID, deviceName, location, type, mqtt_client.client())
     return hubDevice
 
 def start(devicesCurrentSession):
