+++ simuliot/rootfs/starthere/views.py
@@ -81,16 +81,16 @@ def create_session(request):
 def display_session(request):
 	if request.method == 'GET':
 		devices = []
-		req = urllib.request.urlopen('http://127.0.0.1:8088/start-session')
-		if req.status == 200:
-			devices_back = json.loads(urllib.request.urlopen('http://127.0.0.1:8088/retrieve-session').read())
+		#req = urllib.request.urlopen('http://127.0.0.1:8088/start-session')
+		#if req.status == 200:
+		devices_back = json.loads(urllib.request.urlopen('http://127.0.0.1:8088/retrieve-session').read())
 
-			for device in devices_back:
-				new_device = Devices.objects.create_session_device_value(device['id'], device['name'], device['type'], device['location'], device['value'])
-				devices.append(new_device)
-			return render(request, 'display_session.html', {'devices': devices})
-		else:
-			return HttpResponse('Error starting session', status=500)
+		for device in devices_back:
+			new_device = Devices.objects.create_session_device_value(device['id'], device['name'], device['type'], device['location'], device['value'])
+			devices.append(new_device)
+		return render(request, 'display_session.html', {'devices': devices})
+		#else:
+		#	return HttpResponse('Error starting session', status=500)
 
 def terminate_session(request):
 	if request.method == 'GET':
